+++ types/src/block/protocol.rs
@@ -140,6 +140,50 @@ pub fn protocol_parameters() -> ProtocolParameters {
     .unwrap()
 }
 
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+
+    use super::*;
+    use crate::block::{error::dto::DtoError, output::dto::RentStructureDto};
+
+    #[derive(Clone, Debug, PartialEq)]
+    #[cfg_attr(
+        feature = "serde",
+        derive(serde::Serialize, serde::Deserialize),
+        serde(rename_all = "camelCase")
+    )]
+    pub struct ProtocolParametersDto {
+        #[cfg_attr(feature = "serde", serde(rename = "version"))]
+        pub protocol_version: u8,
+        pub network_name: String,
+        pub bech32_hrp: String,
+        pub min_pow_score: u32,
+        pub below_max_depth: u8,
+        pub rent_structure: RentStructureDto,
+        pub token_supply: String,
+    }
+
+    impl TryFrom<ProtocolParametersDto> for ProtocolParameters {
+        type Error = DtoError;
+
+        fn try_from(value: ProtocolParametersDto) -> Result<Self, Self::Error> {
+            Ok(ProtocolParameters::new(
+                value.protocol_version,
+                value.network_name,
+                value.bech32_hrp,
+                value.min_pow_score,
+                value.below_max_depth,
+                value.rent_structure.into(),
+                value
+                    .token_supply
+                    .parse()
+                    .map_err(|_| DtoError::InvalidField("token_supply"))?,
+            )?)
+        }
+    }
+}
+
 #[cfg(feature = "inx")]
 mod inx {
     use packable::PackableExt;