+++ types/CHANGELOG.md
@@ -27,6 +27,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AliasAddress::into_alias_id` method;
 - `Output::{is_treasury, is_basic, is_alias, is_foundry, is_nft}` methods;
 
+### Changed
+
+- Replaced `RentStructureResponse` with `RentStructureDto`;
+- Replaced `ProtocolResponse` with `ProtocolParametersDto`;
+
 ## 1.0.0-rc.2 - 2022-11-09
 
 ### Added